+++ Changelog.md
@@ -1,6 +1,8 @@
 # Change log
 
 ## master (unreleased)
+### Change
+* Change method `get_tests_by_result` for working with array of statuses
 ### Fixes
 * Fix problem with adding result to testrail if test took less than 1 second
 * Fix compatibility with ruby-2.4